# added clean C/C++ keyword lists and use them in `TokenList`