There are various hard-coded lists of these across the code. Some of them are even incomplete, contain non-keywords or attribute keywords to a standard they do not belong to.

This is in preparation so those instances can be incrementally be replaced with these proper global lists.