@firewave 2022-05-16
The interface might not be the best. I will try to replace 1 or 2 of the existing usages locally first to see if it actually makes sense.