@firewave 2023-02-19
Using the static lists instead of building them on-demand in `TokenList` should speed up the tests - especially the scanf/printf ones in `TestIO`.

before:
```
Benchmark 1: bin/testrunner TestIO
  Time (mean ± σ):      2.771 s ±  0.053 s    [User: 2.556 s, System: 0.194 s]
  Range (min … max):    2.689 s …  2.821 s    5 runs
```

after:
```
Benchmark 1: bin/testrunner TestIO
  Time (mean ± σ):      1.967 s ±  0.072 s    [User: 1.806 s, System: 0.161 s]
  Range (min … max):    1.860 s …  2.036 s    5 runs
```